backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
openai>=1.30.0
googlemaps>=4.10.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
